Refactor truth and falsity
