+++ .changeset/early-pumpkins-play.md
@@ -1,0 +1,5 @@
+---
+'tsargp': minor
+---
+
+Updated the code to consider the boolean's truth and falsity names wherever the `enums` attribute is used.
