+++ .changeset/swift-suns-cross.md
@@ -1,0 +1,5 @@
+---
+'@trulysimple/tsargp-docs': patch
+---
+
+Updated the documentation to mention the boolean's truth and falsity names wherever the `enums` attribute is documented.
